requirements.txt
PyYAML~=6.0
torch~=1.11.0
torchvision~=0.12.0
numpy~=1.22.3
scikit-learn~=1.0.2
matplotlib~=3.5.1
seaborn~=0.11.2
pandas~=1.4.1